Accomplish Spring Cloud JPA module
