+++ springcloud-jpa-seata/account-service/pom.xml
@@ -8,10 +8,10 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>user-service</artifactId>
+    <artifactId>account-service</artifactId>
     <packaging>jar</packaging>
 
-    <name>user-service</name>
+    <name>account-service</name>
     <description>Demo project for Spring Boot</description>
 
     <dependencies>
