+++ springcloud-jpa-seata/account-service/src/main/java/io/seata/sample/config/DataSourceConfig.java
@@ -1,0 +1,37 @@
+package io.seata.sample.config;
+
+import com.alibaba.druid.pool.DruidDataSource;
+import io.seata.rm.datasource.DataSourceProxy;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+
+import javax.sql.DataSource;
+
+/**
+ * 数据源配置
+ *
+ * @author HelloWoodes
+ */
+@Configuration
+public class DataSourceConfig {
+
+    @Bean
+    @ConfigurationProperties(prefix = "spring.datasource")
+    public DruidDataSource druidDataSource() {
+        return new DruidDataSource();
+    }
+
+    /**
+     * 需要将 DataSourceProxy 设置为主数据源，否则事务无法回滚
+     *
+     * @param druidDataSource The DruidDataSource
+     * @return The default datasource
+     */
+    @Primary
+    @Bean("dataSource")
+    public DataSource dataSource(DruidDataSource druidDataSource) {
+        return new DataSourceProxy(druidDataSource);
+    }
+}
