+++ springcloud-jpa-seata/account-service/src/main/resources/application.properties
@@ -1,0 +1,6 @@
+spring.application.name=account-service
+server.port=8083
+spring.datasource.url=jdbc:mysql://127.0.0.1:3306/db_account?useSSL=false&serverTimezone=UTC
+spring.datasource.username=root
+spring.datasource.password=123456
+spring.jpa.show-sql=true
