+++ springcloud-jpa-seata/account-service/src/main/resources/file.conf
@@ -20,18 +20,46 @@ transport {
     worker-thread-size = 8
   }
 }
+## transaction log store
+store {
+  ## store mode: file、db
+  mode = "file"
+
+  ## file store
+  file {
+    dir = "sessionStore"
+
+    # branch session size , if exceeded first try compress lockkey, still exceeded throws exceptions
+    max-branch-session-size = 16384
+    # globe session size , if exceeded throws exceptions
+    max-global-session-size = 512
+    # file buffer size , if exceeded allocate new buffer
+    file-write-buffer-cache-size = 16384
+    # when recover batch read size
+    session.reload.read_size = 100
+    # async, sync
+    flush-disk-mode = async
+  }
+
+  ## database store
+  db {
+    driver_class = ""
+    url = ""
+    user = ""
+    password = ""
+  }
+
+}
 service {
   #vgroup->rgroup
-  vgroup_mapping.my_test_tx_group = "default"
+  vgroup_mapping.account-service-fescar-service-group = "default"
   #only support single node
   default.grouplist = "127.0.0.1:8091"
   #degrade current not support
   enableDegrade = false
   #disable
   disable = false
-  disableGlobalTransaction = false
 }
-
 client {
   async.commit.buffer.limit = 10000
   lock {
