+++ springcloud-jpa-seata/account-service/src/main/resources/registry.conf
@@ -1,0 +1,15 @@
+registry {
+  type = "file"
+
+  file {
+    name = "file.conf"
+  }
+}
+
+config {
+  type = "file"
+
+  file {
+    name = "file.conf"
+  }
+}
