+++ springcloud-jpa-seata/business-service/src/main/resources/application.properties
@@ -1,4 +1,2 @@
 spring.application.name=business-service
 server.port=8084
-logging.level.io.seata=debug
-spring.cloud.alibaba.seata.tx-service-group=my_test_tx_group
