+++ springcloud-jpa-seata/order-service/src/main/java/io/seata/sample/entity/Order.java
@@ -1,15 +1,10 @@
 package io.seata.sample.entity;
 
-import java.math.BigDecimal;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
+
+import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "order_tbl")
@@ -20,10 +15,28 @@ public class Order {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+    @Column(name = "user_id")
     private String userId;
+
+    @Column(name = "commodity_code")
     private String commodityCode;
+
+    @Column(name = "money")
     private BigDecimal money;
+
+    @Column(name = "count")
     private Integer count;
+
+    public Order() {
+    }
+
+    public Order(String userId, String commodityCode, BigDecimal money, Integer count) {
+        this.userId = userId;
+        this.commodityCode = commodityCode;
+        this.money = money;
+        this.count = count;
+    }
 
     public Long getId() {
         return id;
