+++ springcloud-jpa-seata/order-service/src/main/java/io/seata/sample/feign/AccountFeignClient.java
@@ -1,10 +1,10 @@
 package io.seata.sample.feign;
 
-import java.math.BigDecimal;
-
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+
+import java.math.BigDecimal;
 
 /**
  * Description：
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RequestParam;
  * @author fangliangsheng
  * @date 2019-04-04
  */
-@FeignClient(name = "user-service", url = "127.0.0.1:8083")
-public interface UserFeignClient {
+@FeignClient(name = "account-service", url = "127.0.0.1:8083")
+public interface AccountFeignClient {
 
     @GetMapping("/debit")
     Boolean debit(@RequestParam("userId") String userId, @RequestParam("money") BigDecimal money);
