+++ springcloud-jpa-seata/order-service/src/main/resources/application.properties
@@ -1,10 +1,5 @@
 spring.application.name=order-service
 server.port=8082
 spring.datasource.url=jdbc:mysql://127.0.0.1:3306/db_order?useSSL=false&serverTimezone=UTC
-spring.datasource.username=mysql
-spring.datasource.password=mysql
-spring.jpa.show-sql:true
-spring.jpa.format-sql:true
-spring.cloud.alibaba.seata.tx-service-group=my_test_tx_group
-logging.level.org.springframework.cloud.alibaba.seata.web=debug
-logging.level.io.seata=debug
+spring.datasource.username=root
+spring.datasource.password=123456
