+++ springcloud-jpa-seata/sql/all_in_one.sql
@@ -1,79 +1,98 @@
-create schema db_account;
-use db_account;
+# Account
+DROP SCHEMA IF EXISTS db_account;
+CREATE SCHEMA db_account;
+USE db_account;
 
-CREATE TABLE `account_tbl` (
-       `id` int(11) NOT NULL AUTO_INCREMENT,
-       `user_id` varchar(255) DEFAULT NULL,
-       `money` int(11) DEFAULT 0,
-       PRIMARY KEY (`id`)
-) ENGINE=InnoDB DEFAULT CHARSET=utf8;
+CREATE TABLE `account_tbl`
+(
+    `id`      INT(11) NOT NULL AUTO_INCREMENT,
+    `user_id` VARCHAR(255) DEFAULT NULL,
+    `money`   INT(11)      DEFAULT 0,
+    PRIMARY KEY (`id`)
+) ENGINE = InnoDB
+  DEFAULT CHARSET = utf8;
 
-INSERT INTO account_tbl (id, user_id, money) VALUES (1, '1001', 10000);
-INSERT INTO account_tbl (id, user_id, money) VALUES (2, '1002', 10000);
+INSERT INTO account_tbl (id, user_id, money)
+VALUES (1, '1001', 10000);
+INSERT INTO account_tbl (id, user_id, money)
+VALUES (2, '1002', 10000);
 
-CREATE TABLE `undo_log` (
-  `id` bigint(20) NOT NULL AUTO_INCREMENT,
-  `branch_id` bigint(20) NOT NULL,
-  `xid` varchar(100) NOT NULL,
-  `rollback_info` longblob NOT NULL,
-  `log_status` int(11) NOT NULL,
-  `log_created` datetime NOT NULL,
-  `log_modified` datetime NOT NULL,
-  `ext` varchar(100) DEFAULT NULL,
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `ux_undo_log` (`xid`,`branch_id`)
-) ENGINE=InnoDB AUTO_INCREMENT=9 DEFAULT CHARSET=utf8;
+CREATE TABLE `undo_log`
+(
+    `id`            BIGINT(20)   NOT NULL AUTO_INCREMENT,
+    `branch_id`     BIGINT(20)   NOT NULL,
+    `xid`           VARCHAR(100) NOT NULL,
+    `rollback_info` LONGBLOB     NOT NULL,
+    `log_status`    INT(11)      NOT NULL,
+    `log_created`   DATETIME     NOT NULL,
+    `log_modified`  DATETIME     NOT NULL,
+    `ext`           VARCHAR(100) DEFAULT NULL,
+    PRIMARY KEY (`id`),
+    UNIQUE KEY `ux_undo_log` (`xid`, `branch_id`)
+) ENGINE = InnoDB
+  DEFAULT CHARSET = utf8;
 
-create schema db_order;
+# Order
+DROP SCHEMA IF EXISTS db_order;
+CREATE SCHEMA db_order;
+USE db_order;
 
-use db_order;
+CREATE TABLE `order_tbl`
+(
+    `id`             INT(11) NOT NULL AUTO_INCREMENT,
+    `user_id`        VARCHAR(255) DEFAULT NULL,
+    `commodity_code` VARCHAR(255) DEFAULT NULL,
+    `count`          INT(11)      DEFAULT '0',
+    `money`          INT(11)      DEFAULT '0',
+    PRIMARY KEY (`id`)
+) ENGINE = InnoDB
+  DEFAULT CHARSET = utf8;
 
-CREATE TABLE `order_tbl` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `user_id` varchar(255) DEFAULT NULL,
-  `commodity_code` varchar(255) DEFAULT NULL,
-  `count` int(11) DEFAULT '0',
-  `money` int(11) DEFAULT '0',
-  PRIMARY KEY (`id`)
-) ENGINE=InnoDB AUTO_INCREMENT=9 DEFAULT CHARSET=utf8;
+CREATE TABLE `undo_log`
+(
+    `id`            BIGINT(20)   NOT NULL AUTO_INCREMENT,
+    `branch_id`     BIGINT(20)   NOT NULL,
+    `xid`           VARCHAR(100) NOT NULL,
+    `rollback_info` LONGBLOB     NOT NULL,
+    `log_status`    INT(11)      NOT NULL,
+    `log_created`   DATETIME     NOT NULL,
+    `log_modified`  DATETIME     NOT NULL,
+    `ext`           VARCHAR(100) DEFAULT NULL,
+    PRIMARY KEY (`id`),
+    UNIQUE KEY `ux_undo_log` (`xid`, `branch_id`)
+) ENGINE = InnoDB
+  DEFAULT CHARSET = utf8;
 
-CREATE TABLE `undo_log` (
-  `id` bigint(20) NOT NULL AUTO_INCREMENT,
-  `branch_id` bigint(20) NOT NULL,
-  `xid` varchar(100) NOT NULL,
-  `rollback_info` longblob NOT NULL,
-  `log_status` int(11) NOT NULL,
-  `log_created` datetime NOT NULL,
-  `log_modified` datetime NOT NULL,
-  `ext` varchar(100) DEFAULT NULL,
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `ux_undo_log` (`xid`,`branch_id`)
-) ENGINE=InnoDB AUTO_INCREMENT=9 DEFAULT CHARSET=utf8;
+# Storage
+DROP SCHEMA IF EXISTS db_storage;
+CREATE SCHEMA db_storage;
+USE db_storage;
 
-create schema db_storage;
-
-use db_storage;
-
-CREATE TABLE `storage_tbl` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
-  `commodity_code` varchar(255) DEFAULT NULL,
-  `count` int(11) DEFAULT '0',
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `commodity_code` (`commodity_code`)
-) ENGINE=InnoDB AUTO_INCREMENT=2 DEFAULT CHARSET=utf8;
+CREATE TABLE `storage_tbl`
+(
+    `id`             INT(11) NOT NULL AUTO_INCREMENT,
+    `commodity_code` VARCHAR(255) DEFAULT NULL,
+    `count`          INT(11)      DEFAULT '0',
+    PRIMARY KEY (`id`),
+    UNIQUE KEY `commodity_code` (`commodity_code`)
+) ENGINE = InnoDB
+  DEFAULT CHARSET = utf8;
 
 
-INSERT INTO storage_tbl (id, commodity_code, count) VALUES (1, '2001', 1000);
+INSERT INTO storage_tbl (id, commodity_code, count)
+VALUES (1, '2001', 1000);
 
-CREATE TABLE `undo_log` (
-  `id` bigint(20) NOT NULL AUTO_INCREMENT,
-  `branch_id` bigint(20) NOT NULL,
-  `xid` varchar(100) NOT NULL,
-  `rollback_info` longblob NOT NULL,
-  `log_status` int(11) NOT NULL,
-  `log_created` datetime NOT NULL,
-  `log_modified` datetime NOT NULL,
-  `ext` varchar(100) DEFAULT NULL,
-  PRIMARY KEY (`id`),
-  UNIQUE KEY `ux_undo_log` (`xid`,`branch_id`)
-) ENGINE=InnoDB AUTO_INCREMENT=9 DEFAULT CHARSET=utf8;
+CREATE TABLE `undo_log`
+(
+    `id`            BIGINT(20)   NOT NULL AUTO_INCREMENT,
+    `branch_id`     BIGINT(20)   NOT NULL,
+    `xid`           VARCHAR(100) NOT NULL,
+    `rollback_info` LONGBLOB     NOT NULL,
+    `log_status`    INT(11)      NOT NULL,
+    `log_created`   DATETIME     NOT NULL,
+    `log_modified`  DATETIME     NOT NULL,
+    `ext`           VARCHAR(100) DEFAULT NULL,
+    PRIMARY KEY (`id`),
+    UNIQUE KEY `ux_undo_log` (`xid`, `branch_id`)
+) ENGINE = InnoDB
+  DEFAULT CHARSET = utf8;
