+++ springcloud-jpa-seata/storage-service/src/main/java/io/seata/sample/StorageApplication.java
@@ -1,13 +1,7 @@
 package io.seata.sample;
 
-import com.alibaba.druid.pool.DruidDataSource;
-
-import io.seata.rm.datasource.DataSourceProxy;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Primary;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
@@ -16,19 +10,6 @@ public class StorageApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(StorageApplication.class, args);
-    }
-
-    @Bean
-    @ConfigurationProperties(prefix = "spring.datasource")
-    public DruidDataSource druidDataSource() {
-        DruidDataSource druidDataSource = new DruidDataSource();
-        return druidDataSource;
-    }
-
-    @Primary
-    @Bean("dataSource")
-    public DataSourceProxy dataSource(DruidDataSource druidDataSource) {
-        return new DataSourceProxy(druidDataSource);
     }
 
 }
