+++ springcloud-jpa-seata/storage-service/src/main/resources/application.properties
@@ -1,10 +1,6 @@
 spring.application.name=storage-service
 server.port=8081
 spring.datasource.url=jdbc:mysql://127.0.0.1:3306/db_storage?useSSL=false&serverTimezone=UTC
-spring.datasource.username=mysql
-spring.datasource.password=mysql
-spring.jpa.show-sql:true
-spring.jpa.format-sql:true
-spring.cloud.alibaba.seata.tx-service-group=my_test_tx_group
-logging.level.org.springframework.cloud.alibaba.seata.web=debug
-logging.level.io.seata=debug
+spring.datasource.username=root
+spring.datasource.password=123456
+spring.jpa.show-sql=true
