+++ springcloud-jpa-seata/storage-service/src/main/resources/file.conf
@@ -22,7 +22,7 @@ transport {
 }
 service {
   #vgroup->rgroup
-  vgroup_mapping.my_test_tx_group = "default"
+  vgroup_mapping.storage-service-fescar-service-group = "default"
   #only support single node
   default.grouplist = "127.0.0.1:8091"
   #degrade current not support
