--- springcloud-jpa-seata/user-service/src/main/java/io/seata/sample/AccountApplication.java
@@ -1,35 +1,0 @@
-package io.seata.sample;
-
-import com.alibaba.druid.pool.DruidDataSource;
-
-import io.seata.rm.datasource.DataSourceProxy;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Primary;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@SpringBootApplication
-@EnableFeignClients
-@EnableJpaRepositories
-public class AccountApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AccountApplication.class, args);
-    }
-
-    @Bean
-    @ConfigurationProperties(prefix = "spring.datasource")
-    public DruidDataSource druidDataSource() {
-        DruidDataSource druidDataSource = new DruidDataSource();
-        return druidDataSource;
-    }
-
-    @Primary
-    @Bean("dataSource")
-    public DataSourceProxy dataSource(DruidDataSource druidDataSource) {
-        return new DataSourceProxy(druidDataSource);
-    }
-}
