--- springcloud-jpa-seata/user-service/src/main/resources/application.properties
@@ -1,9 +1,0 @@
-spring.application.name=account-service
-server.port=8083
-spring.datasource.url=jdbc:mysql://127.0.0.1:3306/db_account?useSSL=false&serverTimezone=UTC
-spring.datasource.username=mysql
-spring.datasource.password=mysql
-spring.jpa.show-sql:true
-spring.jpa.format-sql:true
-logging.level.io.seata=debug
-spring.cloud.alibaba.seata.tx-service-group=my_test_tx_group
