--- springcloud-jpa-seata/user-service/src/main/resources/registry.conf
@@ -1,51 +1,0 @@
-registry {
-  # file 、nacos 、eureka、redis、zk
-  type = "file"
-
-  nacos {
-    serverAddr = "localhost"
-    namespace = "public"
-    cluster = "default"
-  }
-  eureka {
-    serviceUrl = "http://localhost:1001/eureka"
-    application = "default"
-    weight = "1"
-  }
-  redis {
-    serverAddr = "localhost:6381"
-    db = "0"
-  }
-  zk {
-    cluster = "default"
-    serverAddr = "127.0.0.1:2181"
-    session.timeout = 6000
-    connect.timeout = 2000
-  }
-  file {
-    name = "file.conf"
-  }
-}
-
-config {
-  # file、nacos 、apollo、zk
-  type = "file"
-
-  nacos {
-    serverAddr = "localhost"
-    namespace = "public"
-    cluster = "default"
-  }
-  apollo {
-    app.id = "fescar-server"
-    apollo.meta = "http://192.168.1.204:8801"
-  }
-  zk {
-    serverAddr = "127.0.0.1:2181"
-    session.timeout = 6000
-    connect.timeout = 2000
-  }
-  file {
-    name = "file.conf"
-  }
-}
